making_photo_service/requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn==0.30.6
boto3==1.35.24
pandas==2.2.3
matplotlib==3.9.2
pillow==10.4.0